Added new dir hover and renamed to HoverDocumentationProvider. Added new Utils class to get flag url.
