+++ src/main/resources/META-INF/plugin.xml
@@ -33,7 +33,7 @@
                     factoryClass="com.launchdarkly.intellij.toolwindow.FlagToolWindowFactory"/>
 
         <documentationProvider order="first"
-                               implementation="com.launchdarkly.intellij.DocumentationProvider"/>
+                               implementation="com.launchdarkly.intellij.hover.HoverDocumentationProvider"/>
         <!--        <configurationType-->
         <!--                implementation="com.launchdarkly.intellij.coderefs.CodeRefsRunConfigurationType"/>-->
